refactor: Session package 이동 및 Maanger 상속하게 변경
