+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -17,9 +17,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import java.util.function.Consumer;
+import org.apache.catalina.session.HttpSession;
+import org.apache.catalina.session.HttpSessionManger;
 import org.apache.coyote.Processor;
-import org.apache.coyote.session.HttpSession;
-import org.apache.coyote.session.HttpSessionManger;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
